mm: Add PKSM (Anon-page KSM)
This is an improvement upon KSM/UKSM. Some basic data structures and routines are borrowed from ksm.c and uksm.c .

Its new features:
1. Full system scan:
It automatically scans all user processes' anonymous pages. Kernel-user interaction to submit a memory area to KSM is no longer needed.

2. High efficiency for anonymous page detection:
It automatically detects anonymous page is created and freed, It will use a new algorithm and mechanism that it directly handle the anonymous page when it be created/freed by the linux kernel. It is no need to waster CPU time to traversing all of the VMA areas to find valid anonymous pages. KSM/UKSM will waster a lot of CPU time to traversing VMAs to find a valid anonymous page, PKSM no need considering it.

3. Full Zero Page consideration
Now pksmd consider full zero pages as special pages and merge them to an special unswappable pksm zero page.

4. Check page content periodically
Pksm considers unstable page contents are volatile, which are add the a FIFO list. It check page content by a hash value periodically. The default scanning cycles is 20 minutes.

Notes:
(Andrei F. 08/06/2015)

This is a forward port to kernel 3.10.

Tested on Universal7420 platform and is able to merge 150-200k pages resulting in freeing up 600-800mB of physical memory.

Current issues:

- KSM behaviour with swap pages will eventually lead to a BUG() due to bad page. It is currently advised to turn off swapping mechanisms when in conjunction with PKSM.

- ARM architecture has no optimized asm routine for zero page scanning and uses a generic C implementation that may lead to performance and power inefficiencies.

Signed-off-by: Andrei F <luxneb@gmail.com>