regmap: core: Cache all registers by default when cache is enabled
Currently all register maps with a cache need to provide a volatile
callback since the default is to assume all registers are volatile.
This is not sensible if we have a cache so change the default to be
fully cached if a cache is provided.

Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit b92be6fecc9f1e8b927d99c12dad9f9dcd729727)

Change-Id: I7df202ac5f70dde792af9b478da34630294cabe2
Signed-off-by: Ravindra Lokhande <rlokhande@nvidia.com>