video: tegra: host: suspend call prepare_poweroff
Earlier finalize_poweroff was called only as part of save_state() and
therefore it was not automatically part of suspend. In practise the
callback was called as powergating was allowed for devices that
required finalize_poweroff() callback. However, if powergating is
disabled for a device and that device requires finalize_poweroff
callback, it will fail.

This patch makes nvhost to call finalize_poweroff callback during
suspend if powergating is not allowed for the device.

Change-Id: I01dae8dbb36ddcd902bb5ccfa93b1c5adbc6c940
Reviewed-on: http://git-master/r/326698
(cherry picked from commit 29d5142b6bb39ea7f487c02b8369bacdc33035b2)

Signed-off-by: Roger Hsieh <rhsieh@nvidia.com>
Change-Id: Idcf80a59bbc1b2dceb61e1ceafa6d6ae11896ac4
Reviewed-on: http://git-master/r/328893
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Simone Willett <swillett@nvidia.com>