video: tegra: host: Add module_busy_noresume()
Occassionally we want to complete certain operation only if the
device is already on but keep the device on over the time of
operation.

This patch adds function nvhost_module_busy_noresume() that increases
the reference counters of the device without actually causing the
device to be turned on. This prevents device from being powered down
if it is already on, but the device won't be powered if it already
was on.

Bug 1404859

Change-Id: I51b070e09e332a633d72511b1244a8896ce4c5c5
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
(cherry picked from commit b4e76d638723002b62622033c82e17f30f4b5e4a)
Reviewed-on: http://git-master/r/345582
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Michael I Gold <gold@nvidia.com>
Tested-by: Michael I Gold <gold@nvidia.com>