HACK: video: tegra: dsi: disable putting ds in lp mode after commands
tegra_dsi_start_host_cmd_v_blank_video puts the DSI controller
back in LP mode if it was in LP mode when it was called.  This
breaks unblanking the JDI display.  Comment it out for now.

Also add a wait for frame end when the DSI controller is in HS
mode.

Change-Id: I99ef256f8aeb6a432bd58b90e0885cc5e98a48ab
Signed-off-by: Colin Cross <ccross@android.com>