CwMcuSensor: fix memory corruption
kstrndup(buf, PAGE_SIZE, GFP_KERNEL) does not return a buffer
of size PAGE_SIZE, it returns a buffer of size
strnlen(buf, PAGE_SIZE) + 1.  In this case, buf is guaranteed to be
zeroed before it is passed in, so kstrndup will always allocate a
1 byte buffer, and then sprint_data proceeds to write up to 4096 bytes
to it.

Get rid of the kstrndup completely, its useless.  Just scnprintf
directly into the output buffer.  Once the buffer is full,
scnprintf will return buf_remaining - 1, buf_remaining will go
to 1, and future scnprintf calls will return 0 without writing
anything to the buffer.

Bug: 13908204
Change-Id: If1646def885d74e7e409870768d8e2adb88082e4
Signed-off-by: Colin Cross <ccross@android.com>