htc_headset_mgr: fix race condition
Using a flag to track the state of the IRQ is racy.  If the interrupt
fires between enable_irq and setting the flag, the interrupt handler
will not see the new flag status and will not disable the interrupt,
leading to an interrupt storm.

disable_irq/enable_irq are reference counted, so there is no need
to track the state outside the debug UART interrupt.  Simplify the
logic so that when the debug UART interrupt fires, it disables the
headset detect interrupt.  If the headset detect interrupt is
currently running, it will call disable_irq again, and then
enable_irq when the debounce handler finishes, which will still
leave the interrupt disabled.  When the debug UART is removed,
it will re-enable the headset detect interrupt.

Change-Id: I37973055c6b7707e0a20f110d3ee01ee933a42a2
Signed-off-by: Colin Cross <ccross@android.com>