net: wireless: bcmdhd: don't advertise FW_ROAM feature to supplicant
In single channel concurrent mode if we forget AP and again
try to connect to same AP, freq conflict is observed.
If driver advertises FW_ROAM, the supplicant wouldn't send the
BSSID & Freq in the connect command allowing the the driver to
choose the AP to connect to. As freq is not know during connection
it raises a freq conflict. The better approach is to let supplicant
to provide channel info and let FW may roam if needed, so don't
advertise that feature to supplicant.

bug 1498942

Change-Id: If812ce64c958d2f28f88c4459065a08de0aa5d69
Signed-off-by: Manikanta <mmaddireddy@nvidia.com>
Reviewed-on: http://git-master/r/397562
(cherry picked from commit 7bc2df511ee2044e80bbaaf5131d6a92267176a2)
Reviewed-on: http://git-master/r/398725
Reviewed-by: Nagarjuna Kristam <nkristam@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>