power: bq2419x: correctly check for thermal profile data
The number of entry for temp range on DT node is counted
using of calls. If there is no entry then it returns -ve
error. Taking care of -ve returns on the entry count.

This will avoid the wrong error message as:
bq2419x 1-006b: voltage thermal profile is not correct

Change-Id: I573d392123c15241c521c3c3d80f988b8b1f25f7
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/401497
Reviewed-on: http://git-master/r/405641
GVS: Gerrit_Virtual_Submit