arm: tegra:Enable VPR init when CMA is disabled.
VPR init was previously enabled when TLK was disabled. CMA was
supposed to do the allocation and initialization for VPR.
But when CMA is disabled, vpr allocation is done by bootloader and
init is done by kernel. But CMA is disabled on some platforms
using TLK, this was causing VPR allocation failure.

Change-Id: I1d61da4d1698bb2475f1b4897964382b12a5d528
Signed-off-by: Gaurav Sarode <gsarode@nvidia.com>
Reviewed-on: http://git-master/r/403842
Reviewed-by: Prajakta Gudadhe <pgudadhe@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>