cy8c_sar: don't take a wake lock in suspend function
Taking a wake lock in a suspend function will block suspend, even if it
is dropped before reaching the end of the suspend callbacks.  In this
case it is unnecessary.  The wake lock is only used in work queued from
the suspend, resume, and sysfs functions.  None of these can be triggered
after the suspend function has run, so just flush the work in the suspend
function to make sure it is done before finishing going to suspend.

Change-Id: Icde4a85ae98ce3a4b03c37c706f4f7ee1c95d894
Signed-off-by: Colin Cross <ccross@android.com>