htc_headset_pmic: disable wake irq when disabling headset detect
When a serial debug cable is connected the headset detect interrupt
must be disabled to avoid the TX line triggering it.  If it is
left as a wakeup interrupt it will still get marked pending, blocking
suspend until it is cleared.  Disable it as a wakeup interrupt when
the serial debug cable is detected.

Change-Id: I089ad90f5fdae3f3310913959f09ef4e8874e19e
Signed-off-by: Colin Cross <ccross@android.com>