video: tegra: nvmap: Use clean for Denver
On ARMv7 using a full flush was necessary for two reasons. One,
the zeroed pages needed to be written to DRAM or the PoU so that
HW devices could not read the memory out from under the CPU's
cache. And two, the cache lines needed to be invalidated so that
if userspace immediately mapped the memory as something other
than cached they would not get invalid hits in the cache.
However, for denver, the risk of invalid hits in the cache due
to lines being left as valid is no more. Thus we only need to
ensure the zeroed memory is past the PoU. For this a clean will
suffice. Also on denver invalidates are extremely expensive.
Therefor removing any invalidates that are not necessary will
greatly help performance.
Change-Id: Ie952f5db62e1312562ef2ef9c61ba3e37e3efc76
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Signed-off-by: Rich Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/407856
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>