usb: phy: tegra: QC2 speed up charger recogniion
The QC2 Wall Charger needs to be reset only when
we boot or reboot the system.  This is because
the wall charger will have already defaulted to a DCP.
Since the internal logic in the Wall Charger will not
see D+/D- programmed during the time available.

If we just plug in the charger then it will be recognised
in time so no reset is necessary.  Removes 1.5s of delay.

bug 1424733
Reviewed-on: http://git-master/r/433790
(cherry picked from commit b1a51cad99e8ec7e309aa33a786bdc41bc61f6db)
Change-Id: I2b6f698e75f29163e9de4dac3f1b2ef1fb0ec6cc
Signed-off-by: Philip Rakity <prakity@nvidia.com>
Reviewed-on: http://git-master/r/435646
GVS: Gerrit_Virtual_Submit
Reviewed-by: Rakesh Babu Bodla <rbodla@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>