Revert "arm64: mm: align copy_*_user accesses"
This reverts commit 275a80309907a687d21a0f6fe991a0710d3a9a10
(arm64: mm: align copy_*_user accesses), which causes the
zeroing loop to write zeroes over all of memory until the
kernel crashes if the unaligned loop faults.  This is caused
by the zeroing loop expecting to find the upper user buffer
boundary in x4, but the newly added loop stores the bottom
bits of the from address there instead.

Change-Id: I56a43fd2a60e0566af783ed79a62d876b9640914
Signed-off-by: Colin Cross <ccross@android.com>