Revert "Revert "arm64: mm: align copy_*_user accesses""
This reverts commit ea6e1835ed4865eebcd8bf2cbc4404cb4b5f3d89.
The unaligned loop is necessary to support userspace components
that pass Device memory into copy_from_user.  The kernel shouldn't
be giving Device memory to userspace, but until that is fixed
we have to leave this patch in.  The next patch will fix the
issues that caused it to be reverted.

Change-Id: Ifffe453ec9580c6325278d6d5ee9ab5dfd423e8a
Signed-off-by: Colin Cross <ccross@android.com>