ASoC: Tegra: Enable ahub clock before accessing its registers
For compress stream, its front end startup function gets called before
back end startup. We are accessing ahub regsiters in FE startup which is
not getting reflected as clock is not enabled and values are getting
written to regmap cache. This patch enables ahub clock in FE startup
and disables them in FE shutdown.

Bug 16683770

Change-Id: Ia5efa27339805eee46934c239d0bbff40d147c2a
Signed-off-by: Ravindra Lokhande <rlokhande@nvidia.com>