sound: soc: tegra: tegra_rt5677: Boost CPU speed while processing hotword.
Processing the results of the hotword need to happen as fast as possible
on the CPU after the IRQ has been triggered. This helps reduce latency
by pushing the CPU into a high power state temporarily (only 400ms),
then returns the CPU policy to what it was before the IRQ. There doesn't
appear to be any lingering power consumption, I tested with a power
monitor and saw the base consumption drop down to what it was before the
interrupt happened.

Bug: 16729130
Change-Id: I30914d08acd7706447773abea27c2f73cdee731d
Signed-off-by: Chris Thornton <thorntonc@google.com>