sound: soc: codec: rt5677: Tweak VAD to deal with new gain
After changing the way the gain operates to address b/17660951, the
VAD is now a bit too aggressive (since the gain is lower). Bumping the
threshold up from -27 to 0 corrects many instances where you have to
make a loud noise in order for the DSP to start processing audio
first.

Bug: 17660951

Change-Id: Id47a781d1477c84a3734f68c992d14439145c6e4
Signed-off-by: Chris Thornton <thorntonc@google.com>