sound: soc: codec: rt5677: Notify firmware to go to sleep or not
Uses PRI-99 to signal to the firmware if it should go to sleep or not on
the SAD. This is needed to keep the bus alive during headphone playback
so we can reconfigure the bus for recording/stopping headphone playback.

Note that this doesn't actually solve the headphone playback issue yet,
but is a first step towards getting this done.

Bug: 17636802

Change-Id: I5abc7086ec1d55e0feb0e478228f498343c460d9
Signed-off-by: Chris Thornton <thorntonc@google.com>