video: tegra: host: Initialize actmon in poweron
Currently, actmon is initialized when the channel is opened for
the first time and closed when the last user closes the
channel. This leaves actmon uninitialized if we suspend and resume
with having the VIC channel node open.

This patch moves actmon initialization happen as part of
finalize_poweron and deinitialization as part of prepeare_poweroff.
These functions are called when we power on and poweroff off the
device.

Nvidia Bug 1560767
Bug: 18538159

Change-Id: I8972fec60d5213f675dd90fbccf23d9374e3ec9c
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>