sound-control: Adapt configuration for Florida-based platform
On the Florida platform the speaker changes from OUT4 to OUT2 and in the default sound settings is chained with the DSP4 of the digital core which seems to be doing some live DRC via microphone input.

We add the proper pathing to serve this new mixer layout for the speaker.

Because there are issues with the equalizers on the WM5110 which seemingly are very differently configured than on the WM5102, we cascade EQ3 and EQ4 in the speaker path to be able to achieve somewhat acceptable audio quality without fine-tuning the frequencies and gains.

We add a switch to be able to avoid the DSP processing in the speaker path as it seems to be very deterioating to audio quality and shows no benefit.