Revert "ASoC: Run bias level changes for all DAPM contexts in parallel"
This reverts commit 9d0624a740b773d961fe0dc454d748c60d892b71.

Scheduling work queue jobs for bias level changes for all devices
in parallel and then waiting for jobs' completion is resulting in
slower execution (an order of magnitude slower) when the system is
not idle compared to serial execution of all bias level changes.

Stats were measured on a cortex-A15 based system.

Bug: 20226809

Change-Id: I90f685c1bc61964562b7665b05d1f72fc210a7f4
Signed-off-by: Maneet Singh <mmaneetsingh@nvidia.com>